requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
